# backend/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Artifex Gallery API",
    "version": "1.0.0",
    "description": "AI-powered image and video gallery with auto-tagging, NSFW detection, and federation support."
  },
  "servers": [{ "url": "/api" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
  },
  "paths": {
...
    "/admin/stats": { "get": { "tags": ["Admin"], "summary": "System statistics", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Stats" } } } },
    "/admin/users": { "get": { "tags": ["Admin"], "summary": "List all users", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "User list" } } } },
    "/admin/users/{id}/role": { "put": { "tags": ["Admin"], "summary": "Change user role", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Updated" } } } },
    "/admin/users/{id}/disable": { "put": { "tags": ["Admin"], "summary": "Toggle user disabled", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Toggled" } } } },
    "/admin/users/{id}/quota": { "put": { "tags": ["Admin"], "summary": "Set storage quota", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Updated" } } } },
    "/admin/users/{id}/data": { "delete": { "tags": ["Admin"], "summary": "Purge user data", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Purged" } } } },
    "/admin/users/{id}": { "delete": { "tags": ["Admin"], "summary": "Delete user account", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }, { "name": "purge", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "Deleted" } } } },
    "/admin/audit": { "get": { "tags": ["Admin"], "summary": "View audit logs", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "action", "in": "query", "schema": { "type": "string" } }, { "name": "user_id", "in": "query", "schema": { "type": "integer" } }], "responses": { "200": { "description": "Audit logs" } } } },
    "/admin/recalculate-storage": { "post": { "tags": ["Admin"], "summary": "Recalculate storage for all users", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Recalculated" } } } },
...
  }
}